requirements.txt
Flask==3.0.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
